requirements.txt
flask==0.10.1
webdavclient
docker-py